Refactor TsValidationEngine
